pyproject.toml
[project]
name = "event-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "django>=5.2.8",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular[sidecar]>=0.29.0",
    "fastapi>=0.121.2",
    "langchain-chroma>=1.0.0",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.7.1",
    "uvicorn>=0.38.0",
    "whitenoise>=6.11.0",
]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B", "N", "S"]
exclude = ["*/tests/*", "*/migrations/*", "build/"]

[tool.ruff.isort]
known-first-party = ["events", "users", "project", "chat", "blob"]
force-sort-within-sections = true
combine-as-imports = true

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "ruff>=0.14.4",
]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "project.settings"
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py"
]
